trying to fix seg fault

+++ Demo/pipeline.c
@@ -68,7 +68,7 @@ int main (void)
     for(int n = 0; n < num_chunks; n++){
         for(int j = 0; j < c_size; j++){
             for(int k = 0; k < c_size; k++){
-                chunks[n][j][k] = imagePix[n * c_size + j][n * c_size + k];
+                chunks[n][j][k] = (int)imagePix[n * c_size + j][n * c_size + k];
             }
         }
     }
